+++ inlang/source-code/cross-sell/cross-sell-sherlock/src/index.ts
@@ -4,13 +4,30 @@ import {
 	isInWorkspaceRecommendation,
 } from "./recommendation/recommendation.js"
 
-export async function isAdopted(
-	fs: NodeishFilesystem,
+/**
+ * Checks if the Sherlock app is adopted within the workspace recommendations.
+ * @param {Object} args - The arguments object.
+ * @param {NodeishFilesystem} args.fs - The filesystem to use for operations.
+ * @param {string} [args.workingDirectory] - The working directory path. Defaults to the current directory.
+ * @returns {Promise<boolean>} - A promise that resolves to true if the project is adopted, otherwise false.
+ */
+export async function isAdopted(args: {
+	fs: NodeishFilesystem
 	workingDirectory?: string
-): Promise<boolean> {
-	return isInWorkspaceRecommendation(fs, workingDirectory)
+}): Promise<boolean> {
+	return isInWorkspaceRecommendation(args.fs, args.workingDirectory)
 }
 
-export async function add(fs: NodeishFilesystem, workingDirectory?: string): Promise<void> {
-	await addRecommendationToWorkspace(fs, workingDirectory)
+/**
+ * Adds the Sherlock app to the workspace recommendations.
+ * @param {Object} args - The arguments object.
+ * @param {NodeishFilesystem} args.fs - The filesystem to use for operations.
+ * @param {string} [args.workingDirectory] - The working directory path. Defaults to the current directory.
+ * @returns {Promise<void>} - A promise that resolves when the project is successfully added.
+ */
+export async function add(args: {
+	fs: NodeishFilesystem
+	workingDirectory?: string
+}): Promise<void> {
+	await addRecommendationToWorkspace(args.fs, args.workingDirectory)
 }