+++ inlang/source-code/editor/src/pages/@host/@owner/@repository/Layout.tsx
@@ -551,7 +551,14 @@ function ProjectMenu() {
 	const shortenProjectName = (projectPath: string) => {
 		const projectPathArray = projectPath.split("/")
 		if (projectPathArray.length > 3) {
-			return "/" + projectPathArray.at(-3) + "/" + projectPathArray.at(-2) + "/" + projectPathArray.at(-1)
+			return (
+				"/" +
+				projectPathArray.at(-3) +
+				"/" +
+				projectPathArray.at(-2) +
+				"/" +
+				projectPathArray.at(-1)
+			)
 		}
 		return projectPath
 	}