+++ inlang/source-code/ide-extension/src/utilities/import/_import.ts
@@ -30,7 +30,7 @@ const createImport: ImportFunction = async (uri: string) => {
 	// http imports yet like Visual Studio Code.
 
 	const moduleAsText = uri.startsWith("http")
-		? await(await fetch(uri, { agent: httpsAgent })).text()
+		? await (await fetch(uri, { agent: httpsAgent })).text()
 		: await fs.readFile(uri, { encoding: "utf-8" })
 
 	try {