+++ inlang/source-code/paraglide/paraglide-js-sveltekit-old/src/adapter-sveltekit/examples/lang-in-slug-static/src/routes/+layout.ts
@@ -1,8 +1,0 @@
-import { initRootLayoutLoadWrapper } from "@inlang/paraglide-js-sveltekit/adapter-sveltekit/shared"
-import type { LayoutLoad } from "./$types.js"
-
-export const load = initRootLayoutLoadWrapper<LayoutLoad>({}).use(async ({ data }, { i }) => {
-	console.info("+layout.ts", i("welcome"))
-
-	return { ...data, "+layout.ts": Math.random() }
-})