+++ inlang/source-code/paraglide/paraglide-js-sveltekit-old/src/adapter-sveltekit/examples/lang-in-slug/src/routes/[lang]/+layout.ts
@@ -1,6 +1,0 @@
-import { initLoadWrapper } from "@inlang/paraglide-js-sveltekit/adapter-sveltekit/shared"
-import type { LayoutLoad } from "./$types.js"
-
-export const load = initLoadWrapper<LayoutLoad>().use(async (_, { i }) => {
-	console.info("[lang]/+layout.ts", i("welcome"))
-})