+++ inlang/source-code/paraglide/paraglide-js-sveltekit-old/src/utils.ts
@@ -1,2 +1,0 @@
-export const logDeprecation = (oldName: string, newName: string) =>
-	console.warn(`[INLANG:DEPRECATED] '${oldName}' is deprecated. Use '${newName}' instead.`)