+++ inlang/source-code/settings-component/src/stories/inlang-settings.ts
@@ -1,22 +1,43 @@
 import { html, LitElement, css } from "lit"
 import { customElement, property, state } from "lit/decorators.js"
 import { baseStyling } from "../styling/base.js"
-import { type InlangProject, ProjectSettings, InlangModule } from "@inlang/sdk"
-import { Task } from "@lit/task"
-import { setBasePath } from "@shoelace-style/shoelace/dist/utilities/base-path.js"
-setBasePath("./../../node_modules/@shoelace-style/shoelace/dist")
+import {
+	type InlangModule,
+	ProjectSettings,
+	type InlangProject,
+	type InstalledPlugin,
+	type InstalledMessageLintRule,
+} from "@inlang/sdk"
 
-import "./input-fields/simple-input.js"
+import "./input-fields/general-input.js"
+
+import SlSelect from "@shoelace-style/shoelace/dist/components/select/select.component.js"
+import SLOption from "@shoelace-style/shoelace/dist/components/option/option.component.js"
+import SlInput from "@shoelace-style/shoelace/dist/components/input/input.component.js"
+import SlButton from "@shoelace-style/shoelace/dist/components/button/button.component.js"
+import SlCheckbox from "@shoelace-style/shoelace/dist/components/checkbox/checkbox.component.js"
+import SlIconButton from "@shoelace-style/shoelace/dist/components/icon-button/icon-button.component.js"
+import SlIcon from "@shoelace-style/shoelace/dist/components/icon/icon.component.js"
+
+import { setBasePath } from "@shoelace-style/shoelace/dist/utilities/base-path.js"
+setBasePath("./../../../../node_modules/@shoelace-style/shoelace/dist")
+
+// in case an app defines it's own set of shoelace components, prevent double registering
+if (!customElements.get("sl-select")) customElements.define("sl-select", SlSelect)
+if (!customElements.get("sl-option")) customElements.define("sl-option", SLOption)
+if (!customElements.get("sl-input")) customElements.define("sl-input", SlInput)
+if (!customElements.get("sl-button")) customElements.define("sl-button", SlButton)
+if (!customElements.get("sl-checkbox")) customElements.define("sl-checkbox", SlCheckbox)
+if (!customElements.get("sl-icon-button")) customElements.define("sl-icon-button", SlIconButton)
+if (!customElements.get("sl-icon")) customElements.define("sl-icon", SlIcon)
 
 @customElement("inlang-settings")
-export class InlangSettings extends LitElement {
+export default class InlangSettings extends LitElement {
 	static override styles = [
 		baseStyling,
 		css`
 			h3 {
 				margin: 0;
-				padding-bottom: 24px;
-				border-bottom: 1px solid var(--sl-panel-border-color);
 			}
 			.container {
 				position: relative;
@@ -27,7 +48,7 @@ export class InlangSettings extends LitElement {
 			.module-container {
 				display: flex;
 				flex-direction: column;
-				gap: 24px;
+				gap: 16px;
 			}
 			.hover-bar-container {
 				width: 100%;
@@ -59,14 +80,11 @@ export class InlangSettings extends LitElement {
 		`,
 	]
 
-	@property()
-	inlangProject: ReturnType<InlangProject["settings"]> | undefined = undefined
-
-	@property()
-	onSaveProject?: (project: ReturnType<InlangProject["settings"]>) => void = () => {}
+	@property({ attribute: false })
+	project: InlangProject = {} as InlangProject
 
 	@state()
-	private _project: ReturnType<InlangProject["settings"]> | undefined = undefined
+	private _newSettings: ProjectSettings | undefined = undefined
 
 	@state()
 	private _unsavedChanges: boolean = false
@@ -74,8 +92,8 @@ export class InlangSettings extends LitElement {
 	override async firstUpdated() {
 		await this.updateComplete
 
-		if (this.inlangProject) {
-			this._project = JSON.parse(JSON.stringify(this.inlangProject))
+		if (this.project?.settings()) {
+			this._newSettings = JSON.parse(JSON.stringify(this.project.settings()))
 		}
 	}
 
@@ -86,21 +104,21 @@ export class InlangSettings extends LitElement {
 		moduleId?: InlangModule["default"]["id"]
 	) => {
 		//update state object
-		if (this._project && moduleId) {
-			this._project = {
-				...this._project,
+		if (this._newSettings && moduleId) {
+			this._newSettings = {
+				...this._newSettings,
 				[moduleId]: {
-					...this._project[moduleId],
+					...this._newSettings[moduleId],
 					[property]: value,
 				},
 			}
-		} else if (this._project) {
-			this._project = {
-				...this._project,
+		} else if (this._newSettings) {
+			this._newSettings = {
+				...this._newSettings,
 				[property]: value,
 			}
 		}
-		if (JSON.stringify(this.inlangProject) !== JSON.stringify(this._project)) {
+		if (JSON.stringify(this.project?.settings()) !== JSON.stringify(this._newSettings)) {
 			this._unsavedChanges = true
 		} else {
 			this._unsavedChanges = false
@@ -108,116 +126,123 @@ export class InlangSettings extends LitElement {
 	}
 
 	_revertChanges = () => {
-		if (this.inlangProject) {
-			this._project = JSON.parse(JSON.stringify(this.inlangProject))
+		if (this.project?.settings()) {
+			this._newSettings = JSON.parse(JSON.stringify(this.project?.settings()))
 		}
 		this._unsavedChanges = false
 	}
 
 	_saveChanges = () => {
-		if (this._project && this.onSaveProject) {
-			this.onSaveProject(this._project)
+		if (this._newSettings) {
+			this.project?.setSettings(this._newSettings)
 		}
 		this._unsavedChanges = false
 	}
 
-	private _projectProperties = new Task(this, {
-		task: async ([inlangProject]) => {
-			if (!inlangProject) throw new Error("No inlang project")
+	private get _projectSettingProperties(): Record<
+		InlangModule["default"]["id"] | "internal",
+		{
+			meta?: InstalledPlugin | InstalledMessageLintRule
+			schema?: Record<string, Record<string, unknown>>
+		}
+	> {
+		const _project = this.project
+		if (!_project) throw new Error("No inlang project")
+		if (!_project.settings()) throw new Error("No inlang project settings")
 
-			const generalSchema: Record<
-				InlangModule["default"]["id"] | "internal",
-				{ meta?: InlangModule["default"]; schema?: Record<string, Record<string, unknown>> }
-			> = { internal: { schema: ProjectSettings.allOf[0] } }
+		const generalSchema: Record<
+			InlangModule["default"]["id"] | "internal",
+			{
+				meta?: InstalledPlugin | InstalledMessageLintRule
+				schema?: Record<string, Record<string, unknown>>
+			}
+		> = { internal: { schema: ProjectSettings.allOf[0] } }
 
-			for (const module of inlangProject.modules) {
-				try {
-					const plugin = await import(module)
-					if (plugin.default) {
-						generalSchema[plugin.default.id] = {
-							schema: plugin.default.settingsSchema,
-							meta: plugin.default,
-						}
-					}
-				} catch (e) {
-					console.error(e)
+		for (const plugin of _project.installed.plugins()) {
+			if (plugin.settingsSchema) {
+				generalSchema[plugin.id] = {
+					schema: plugin.settingsSchema,
+					meta: plugin,
 				}
 			}
-			return generalSchema
-		},
-		args: () => [this.inlangProject],
-	})
+		}
+		for (const lintRule of _project.installed.messageLintRules()) {
+			if (lintRule.settingsSchema) {
+				generalSchema[lintRule.id] = {
+					schema: lintRule.settingsSchema,
+					meta: lintRule,
+				}
+			}
+		}
+
+		return generalSchema
+	}
 
 	override render() {
-		return this._projectProperties.render({
-			pending: () => html`<div>Loading...</div>`,
-			complete: (properties) =>
-				html` <div class="container">
-					${Object.entries(properties).map(([key, value]) => {
-						return value.schema?.properties && this._project
-							? html`<div class="module-container">
-									<h3>
-										${(value.meta as { displayName?: { en: string } })?.displayName?.en || key}
-									</h3>
-									${Object.entries(value.schema.properties).map(([property, schema]) => {
-										if (property === "$schema" || property === "modules") return undefined
-										return key === "internal"
-											? html`
-													<simple-input
-														.property=${property}
-														.modules=${properties}
-														.value=${this._project?.[property as keyof typeof this._project]}
-														.schema=${schema}
-														.handleInlangProjectChange=${this.handleInlangProjectChange}
-													></simple-input>
-											  `
-											: html`
-													<simple-input
-														.property=${property}
-														.value=${
-															// @ts-ignore
-															this._project?.[key]?.[property]
-														}
-														.schema=${schema}
-														.moduleId=${key}
-														.handleInlangProjectChange=${this.handleInlangProjectChange}
-													></simple-input>
-											  `
-									})}
-							  </div>`
-							: undefined
-					})}
-					${this._unsavedChanges
-						? html`<div class="hover-bar-container">
-								<div class="hover-bar">
-									<p class="hover-bar-text">Attention, you have unsaved changes.</p>
-									<div>
-										<sl-button
-											size="medium"
-											@click=${() => {
-												this._revertChanges()
-											}}
-											varaint="default"
-										>
-											Cancel
-										</sl-button>
-										<sl-button
-											size="medium"
-											@click=${() => {
-												this._saveChanges()
-											}}
-											variant="primary"
-										>
-											Save Changes
-										</sl-button>
-									</div>
-								</div>
-						  </div>`
-						: html``}
-				</div>`,
-
-			error: (e) => html`<p>Error: ${e}</p>`,
-		})
+		return html` <div class="container">
+			${Object.entries(this._projectSettingProperties).map(([key, value]) => {
+				return value.schema?.properties && this._newSettings
+					? html`<div class="module-container">
+							${value.meta &&
+							(value.meta?.displayName as { en: string }).en &&
+							html`<h3>${value.meta && (value.meta?.displayName as { en: string }).en}</h3>`}
+							${Object.entries(value.schema.properties).map(([property, schema]) => {
+								if (property === "$schema" || property === "modules" || property === "experimental")
+									return undefined
+								return key === "internal"
+									? html`
+											<general-input
+												.property=${property}
+												.modules=${this.project.installed.messageLintRules() || []}
+												.value=${this._newSettings?.[property as keyof typeof this._newSettings]}
+												.schema=${schema}
+												.handleInlangProjectChange=${this.handleInlangProjectChange}
+											></general-input>
+									  `
+									: html`
+											<general-input
+												.property=${property}
+												.value=${
+													// @ts-ignore
+													this._newSettings?.[key]?.[property]
+												}
+												.schema=${schema}
+												.moduleId=${key}
+												.handleInlangProjectChange=${this.handleInlangProjectChange}
+											></general-input>
+									  `
+							})}
+					  </div>`
+					: undefined
+			})}
+			${this._unsavedChanges
+				? html`<div class="hover-bar-container">
+						<div class="hover-bar">
+							<p class="hover-bar-text">Attention, you have unsaved changes.</p>
+							<div>
+								<sl-button
+									size="medium"
+									@click=${() => {
+										this._revertChanges()
+									}}
+									varaint="default"
+								>
+									Cancel
+								</sl-button>
+								<sl-button
+									size="medium"
+									@click=${() => {
+										this._saveChanges()
+									}}
+									variant="primary"
+								>
+									Save Changes
+								</sl-button>
+							</div>
+						</div>
+				  </div>`
+				: html``}
+		</div>`
 	}
 }