+++ inlang/source-code/github-lint-action/CHANGELOG.md
@@ -1,5 +1,53 @@
 # @inlang/github-lint-action
 
+## 0.2.14
+
+### Patch Changes
+
+- Updated dependencies [d1b361e]
+  - @lix-js/client@1.1.0
+  - @inlang/sdk@0.28.2
+
+## 0.2.13
+
+### Patch Changes
+
+- Updated dependencies [03fa6f2]
+  - @lix-js/client@1.0.0
+  - @lix-js/fs@1.0.0
+  - @inlang/sdk@0.28.1
+
+## 0.2.12
+
+### Patch Changes
+
+- Updated dependencies [1e43ae4]
+  - @inlang/sdk@0.28.0
+
+## 0.2.11
+
+### Patch Changes
+
+- e169f29: Remove project not found comment
+
+## 0.2.10
+
+### Patch Changes
+
+- 033a6cc: Fix project not found error
+
+## 0.2.9
+
+### Patch Changes
+
+- 54d51b7: Handle project not found in head repo
+
+## 0.2.8
+
+### Patch Changes
+
+- 05d1914: set required node version down to v18
+
 ## 0.2.5
 
 ### Patch Changes