+++ inlang/source-code/paraglide/paraglide-js-adapter-next/examples/pages/CHANGELOG.md
@@ -1,5 +1,50 @@
 # @inlang/paraglide-js-adapter-next-example-pages
 
+## 0.2.10
+
+### Patch Changes
+
+- Updated dependencies [7dce581]
+  - @inlang/paraglide-js@1.3.5
+
+## 0.2.9
+
+### Patch Changes
+
+- Updated dependencies [2a42b6e]
+  - @inlang/paraglide-js@1.3.4
+
+## 0.2.8
+
+### Patch Changes
+
+- Updated dependencies [7115a13]
+- Updated dependencies [0dc1be3]
+- Updated dependencies [124435c]
+  - @inlang/paraglide-js-adapter-next@1.1.0
+
+## 0.2.7
+
+### Patch Changes
+
+- Updated dependencies [11f0e18]
+- Updated dependencies [10e3c28]
+  - @inlang/paraglide-js@1.3.3
+
+## 0.2.6
+
+### Patch Changes
+
+- Updated dependencies [45975c0]
+  - @inlang/paraglide-js@1.3.2
+
+## 0.2.5
+
+### Patch Changes
+
+- Updated dependencies [afaaffa]
+  - @inlang/paraglide-js@1.3.1
+
 ## 0.2.4
 
 ### Patch Changes