+++ inlang/source-code/paraglide/paraglide-js/CHANGELOG.md
@@ -1,5 +1,36 @@
 # @inlang/paraglide-js
 
+## 1.3.5
+
+### Patch Changes
+
+- 7dce581: fix `openRepository` crash in non-git environments
+
+## 1.3.4
+
+### Patch Changes
+
+- 2a42b6e: bump `@lix-js/client` dependency
+
+## 1.3.3
+
+### Patch Changes
+
+- 11f0e18: Update dependency
+- 10e3c28: fix JSDoc annotations on message aliases
+
+## 1.3.2
+
+### Patch Changes
+
+- 45975c0: Fail gracefully if adding `.vscode` folder fails
+
+## 1.3.1
+
+### Patch Changes
+
+- afaaffa: Bundle `@inlang/cross-sell-sherlock`. This is used to promot _once_ during `paraglide-js init` to ask if you want to install the Sherlock vscode extension
+
 ## 1.3.0
 
 ### Minor Changes